requirements.txt
flask==2.0.1
Werkzeug==2.0.1
gunicorn==20.1.0
flask-mail==0.9.1
python-dotenv==0.19.0
